requirements.txt
fastapi>=0.110.0,<0.111.0
uvicorn>=0.29.0,<0.30.0
pdfplumber>=0.11.0,<0.12.0
pymupdf>=1.24.1,<1.25.0
requests>=2.31.0,<2.32.0
python-multipart>=0.0.9,<0.1.0
aiofiles>=24.1.0,<25.0.0
python-dotenv>=1.0.1,<1.1.0
numpy>=1.26.0,<2.0.0
protobuf>=4.21.6,<5.0.0  # For grpcio-status compatibility
httpx>=0.27.0,<0.28.0  # For unstructured-client
tiktoken>=0.7.0,<1.0.0  # For langchain-openai
typer>=0.12.3,<0.13.0  # Added for fastapi-cli
fastapi-cli>=0.0.4,<0.1.0  # Added explicitly
tensorflow>=2.16.0,<3.0.0  # Added for tf-keras
tf-keras>=2.19.0,<2.20.0  # Added explicitly